pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["х"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.13.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
